+++ apps/dialtone-documentation/docs/.vuepress/theme/layouts/Layout.vue
@@ -3,7 +3,7 @@
     :fixed="false"
     :header-sticky="true"
     header-class="d-ol-none"
-    sidebar-class="dialtone-sidebar lg:d-d-none d-ol-none"
+    sidebar-class="dialtone-sidebar d-d-none lg:d-d-block d-ol-none"
     footer-class="d-text-right d-ol-none"
     content-class="d-ol-none"
   >
