+++ apps/dialtone-documentation/docs/components/icon.md
@@ -44,13 +44,13 @@ import { DtIcon } from '@dialpad/dialtone'
 
 ## Changing Sizes
 
-<div class="d-d-grid d-g24 d-g-cols3 md:d-g-cols1">
+<div class="d-d-grid d-g24 d-g-cols1 md:d-g-cols3">
 
 Adjust the size using the `size` prop. Note that sizes 600, 700, and 800 are exclusively for devices.
 
 <div class="d-gc2">
 <code-well-header>
-  <div class="d-d-grid d-g24 d-g-cols2 md:d-g-cols1 d-w100p">
+  <div class="d-d-grid d-g24 d-g-cols1 md:d-g-cols2 d-w100p">
     <div class="d-fl-center">
       <dt-icon-settings :size="selectedSize" />
     </div>
@@ -67,13 +67,13 @@ Adjust the size using the `size` prop. Note that sizes 600, 700, and 800 are exc
 
 ## Changing Color
 
-<div class="d-d-grid d-g24 d-g-cols3 md:d-g-cols1">
+<div class="d-d-grid d-g24 d-g-cols1 md:d-g-cols3">
 
 The icon's color inherits from the parent's foreground color.
 
 <div class="d-gc2">
 <code-well-header>
-<div class="d-d-grid d-g24 d-g-cols2 md:d-g-cols1 d-w100p">
+<div class="d-d-grid d-g24 d-g-cols1 md:d-g-cols2 d-w100p">
   <div class="d-fl-center">
     <dt-stack :class="selectedColor" direction="row" as="div" gap="300">
       <dt-icon-settings size="300" />
@@ -94,7 +94,7 @@ The icon's color inherits from the parent's foreground color.
 </div>
 </div>
 
-<div class="d-d-grid d-g24 d-g-cols3 md:d-g-cols1">
+<div class="d-d-grid d-g24 d-g-cols1 md:d-g-cols3">
 
 When setting the color of an icon take these into considaration:
 
@@ -139,13 +139,13 @@ When setting the color of an icon take these into considaration:
 
 ## Icon and Text Alignment
 
-<div class="d-d-grid d-g24 d-g-cols3 md:d-g-cols1">
+<div class="d-d-grid d-g24 d-g-cols1 md:d-g-cols3">
 
 We encourage utilizing the [Stack component](/components/stack.md) for aligning elements both horizontally and vertically.
 
 <div class="d-gc2">
 <code-well-header>
-  <div class="d-d-grid d-g24 d-g-cols2 md:d-g-cols1 d-w100p">
+  <div class="d-d-grid d-g24 d-g-cols1 md:d-g-cols2 d-w100p">
     <div class="d-fl-center">
       <dt-stack :direction="selectedDirection" class="d-fl-center" gap="300">
       <dt-icon-settings size="300" />
