+++ apps/dialtone-documentation/docs/design/space/index.md
@@ -24,7 +24,7 @@ When setting the scale of an element, use [Size](/design/size/index.md) units.
 
 ### Setting a Specific Spacing
 
-<div class="d-d-grid d-g24 d-g-cols3 md:d-g-cols1">
+<div class="d-d-grid d-g24 d-g-cols1 md:d-g-cols3">
 
 <div>
 
@@ -72,7 +72,7 @@ When setting the gap between elements, utilize [Space Tokens](#tokens) for the g
 
 ### Stacking
 
-<div class="d-d-grid d-g24 d-g-cols3 md:d-g-cols1">
+<div class="d-d-grid d-g24 d-g-cols1 md:d-g-cols3">
 <div>
 
 When adding space between elements, both vertically and horizontally, we can use either `space-between` or `gap` to define the space that separates each element. The value is set as a variable in Figma and as a token in code.
@@ -99,7 +99,7 @@ When adding space between elements, both vertically and horizontally, we can use
 
 ### Padding
 
-<div class="d-d-grid d-g24 d-g-cols3 md:d-g-cols1">
+<div class="d-d-grid d-g24 d-g-cols1 md:d-g-cols3">
 <div>
 
 When adding space around an element, use `padding` to create breathing room.
@@ -120,7 +120,7 @@ Avoid using `margin`, which adds space outside the element and can affect the la
 
 ### Grouping
 
-<div class="d-d-grid d-g16 d-g-cols3 md:d-g-cols1" >
+<div class="d-d-grid d-g16 d-g-cols1 md:d-g-cols3" >
   <div>
 <dt-stack direction="column" gap="400">
 <p>Group and adjust the distance between elements to guide users understanding of their relationships. Closer elements appear more connected, while larger gaps indicate less association.</p>
@@ -137,7 +137,7 @@ Avoid using `margin`, which adds space outside the element and can affect the la
 
 ### Guiding the Eye
 
-<div class="d-d-grid d-g16 d-g-cols3 md:d-g-cols1" >
+<div class="d-d-grid d-g16 d-g-cols1 md:d-g-cols3" >
   <div>
     <dt-stack direction="column" gap="400">
       <div>
