+++ apps/dialtone-documentation/docs/utilities/responsive/breakpoints.md
@@ -13,18 +13,18 @@ This plugin takes the breakpoints and the classes you need to have responsive va
 
 <code-well-header bgclass='d-bgc-purple-100'>
   <div class="d-fl-center d-w100p d-m8 d-p16 d-bgc-purple-300 d-bar4 d-fs-300 d-fw-bold d-ta-center">This is visible on all screens</div>
-  <div class="d-fl-center d-w100p d-m8 d-p16 d-bgc-purple-300 d-bar4 d-fs-300 d-fw-bold d-ta-center d-d-none xl:d-d-block">This is visible on extra large screens</div>
-  <div class="d-fl-center d-w100p d-m8 d-p16 d-bgc-purple-300 d-bar4 d-fs-300 d-fw-bold d-ta-center d-d-none lg:d-d-block">This is visible on large screens</div>
-  <div class="d-fl-center d-w100p d-m8 d-p16 d-bgc-purple-300 d-bar4 d-fs-300 d-fw-bold d-ta-center d-d-none md:d-d-block">This is visible on medium screens</div>
-  <div class="d-fl-center d-w100p d-m8 d-p16 d-bgc-purple-300 d-bar4 d-fs-300 d-fw-bold d-ta-center d-d-none sm:d-d-block">This is visible on small screens</div>
+  <div class="d-fl-center d-w100p d-m8 d-p16 d-bgc-purple-300 d-bar4 d-fs-300 d-fw-bold d-ta-center d-d-none sm:d-d-block">This is visible on small screens and above</div>
+  <div class="d-fl-center d-w100p d-m8 d-p16 d-bgc-purple-300 d-bar4 d-fs-300 d-fw-bold d-ta-center d-d-none md:d-d-block">This is visible on medium screens and above</div>
+  <div class="d-fl-center d-w100p d-m8 d-p16 d-bgc-purple-300 d-bar4 d-fs-300 d-fw-bold d-ta-center d-d-none lg:d-d-block">This is visible on large screens and above</div>
+  <div class="d-fl-center d-w100p d-m8 d-p16 d-bgc-purple-300 d-bar4 d-fs-300 d-fw-bold d-ta-center d-d-none xl:d-d-block">This is visible on extra large screens and above</div>
 </code-well-header>
 
 ```html
 <div>...</div>
-<div class="d-d-none xl:d-d-block">...</div>
-<div class="d-d-none lg:d-d-block">...</div>
-<div class="d-d-none md:d-d-block">...</div>
 <div class="d-d-none sm:d-d-block">...</div>
+<div class="d-d-none md:d-d-block">...</div>
+<div class="d-d-none lg:d-d-block">...</div>
+<div class="d-d-none xl:d-d-block">...</div>
 ```
 
 ## Classes
@@ -43,24 +43,24 @@ To help keep prefixes concise, we use abbreviations. This syntax is used consist
       </thead>
       <tbody>
         <tr>
-          <th scope="row" class="d-code--sm d-docsite-code">.xl:{class}</th>
-          <td class="d-code--sm">max-width: 1264px</td>
-          <td class="d-code--sm">The class is applied on extra large browser widths and below.</td>
-        </tr>
-        <tr>
-          <th scope="row" class="d-code--sm d-docsite-code">.lg:{class}</th>
-          <td class="d-code--sm">max-width: 980px</td>
-          <td class="d-code--sm">The class is applied on large browser widths and below.</td>
+          <th scope="row" class="d-code--sm d-docsite-code">.sm:{class}</th>
+          <td class="d-code--sm">min-width: 480px</td>
+          <td class="d-code--sm">The class is applied on small browser widths and above.</td>
         </tr>
         <tr>
           <th scope="row" class="d-code--sm d-docsite-code">.md:{class}</th>
-          <td class="d-code--sm">max-width: 640px</td>
-          <td class="d-code--sm">The class is applied on medium browser widths and below.</td>
+          <td class="d-code--sm">min-width: 640px</td>
+          <td class="d-code--sm">The class is applied on medium browser widths and above.</td>
         </tr>
         <tr>
-          <th scope="row" class="d-code--sm d-docsite-code">.sm:{class}</th>
-          <td class="d-code--sm">max-width: 480px</td>
-          <td class="d-code--sm">The class is applied on small browser widths and below.</td>
+          <th scope="row" class="d-code--sm d-docsite-code">.lg:{class}</th>
+          <td class="d-code--sm">min-width: 980px</td>
+          <td class="d-code--sm">The class is applied on large browser widths and above.</td>
+        </tr>
+        <tr>
+          <th scope="row" class="d-code--sm d-docsite-code">.xl:{class}</th>
+          <td class="d-code--sm">min-width: 1264px</td>
+          <td class="d-code--sm">The class is applied on extra large browser widths and above.</td>
         </tr>
       </tbody>
     </table>
