+++ apps/dialtone-documentation/docs/utilities/typography/text-align.md
@@ -8,7 +8,7 @@ description: Utilities for controlling an element's text alignment.
 Use `d-ta-{n}` to change an element's text alignment.
 
 <code-well-header class="d-p24 d-bgc-purple-100 d-bgo50 d-w100p d-hmn102 d-of-auto" custom>
-  <div class="d-d-grid d-g16 d-ai-center lg:d-fs-100" style="grid-template-columns: auto 1fr">
+  <div class="d-d-grid d-g16 d-ai-center d-fs-100 lg:d-fs-200" style="grid-template-columns: auto 1fr">
     <div class="d-code--sm d-fc-blue-500 d-ws-nowrap">.d-ta-left</div>
     <div><p class="d-ta-left d-pr12">The quick brown fox jumps over the lazy dog.</p></div>
     <div class="d-code--sm d-fc-blue-500 d-ws-nowrap">.d-ta-center</div>
