+++ apps/dialtone-documentation/docs/utilities/typography/word-wrap.md
@@ -8,7 +8,7 @@ description: Utilities for controlling the way words wrap within an element. Gen
 Use `d-ww-normal` to break words only at allowed break points.
 
 <code-well-header class="d-fl-center d-p24 d-bgc-purple-100 d-bgo50 d-w100p d-hmn102" custom>
-  <div class="d-bgc-purple-200 d-py8 d-px16 d-bar8 lg:d-w216 d-w332">
+  <div class="d-bgc-purple-200 d-py8 d-px16 d-bar8 d-w216 lg:d-w332">
     <p class="d-fs-200 d-ww-normal">Here's an example sentence to show how word-wrap works. Thisisasignlewordtodenotethedifferencebetweenthedifferentwaytowrapaword. Vivamus ullamcorperatduiaultrices eu lobortis nulla, sed vulputate orci. 这是一个中文例句，以举例说明断字的工作方式。単語分割の動作の例を示す日本語のサンプル文は次のとおりです。다음은 단어 분리 작동 방식의 예를 제공하는 한국어 샘플 문장입니다.</p>
   </div>
 </code-well-header>
@@ -22,7 +22,7 @@ Use `d-ww-normal` to break words only at allowed break points.
 Use `d-ww-break-word` to allow unbreakable words to be broken. Is a more conservative approach than [`d-wb-break-all`](./word-break.md#break-all) and will only break long words that do not fit the container.
 
 <code-well-header class="d-fl-center d-p24 d-bgc-green-100 d-bgo50 d-w100p d-hmn102" custom>
-  <div class="d-bgc-green-100 d-py8 d-px16 d-bar8 lg:d-w216 d-w332">
+  <div class="d-bgc-green-100 d-py8 d-px16 d-bar8 d-w216 lg:d-w332">
     <p class="d-fs-200 d-ww-break-word">Here's an example sentence to show how word-wrap works. Thisisasignlewordtodenotethedifferencebetweenthedifferentwaytowrapaword. Vivamus ullamcorperatduiaultrices eu lobortis nulla, sed vulputate orci. 这是一个中文例句，以举例说明断字的工作方式。単語分割の動作の例を示す日本語のサンプル文は次のとおりです。다음은 단어 분리 작동 방식의 예를 제공하는 한국어 샘플 문장입니다.</p>
   </div>
 </code-well-header>
@@ -36,7 +36,7 @@ Use `d-ww-break-word` to allow unbreakable words to be broken. Is a more conserv
 Use `d-ww-initial`to set this property to its default value.
 
 <code-well-header class="d-fl-center d-p24 d-bgc-magenta-100 d-bgo50 d-w100p d-hmn102" custom>
-  <div class="d-bgc-magenta-100 d-py8 d-px16 d-bar8 lg:d-w216 d-w332">
+  <div class="d-bgc-magenta-100 d-py8 d-px16 d-bar8 d-w216 lg:d-w332">
     <p class="d-fs-200 d-ww-initial">Here's an example sentence to show how word-wrap works. Thisisasignlewordtodenotethedifferencebetweenthedifferentwaytowrapaword. Vivamus ullamcorperatduiaultrices eu lobortis nulla, sed vulputate orci. 这是一个中文例句，以举例说明断字的工作方式。単語分割の動作の例を示す日本語のサンプル文は次のとおりです。다음은 단어 분리 작동 방식의 예를 제공하는 한국어 샘플 문장입니다.</p>
   </div>
 </code-well-header>
@@ -50,7 +50,7 @@ Use `d-ww-initial`to set this property to its default value.
 Use `d-ww-inherit` to inherit this property from its parent element.
 
 <code-well-header class="d-fl-center d-p24 d-bgc-magenta-100 d-bgo50 d-w100p d-hmn102" custom>
-  <div class="d-bgc-magenta-100 d-py8 d-px16 d-bar8 lg:d-w216 d-w332">
+  <div class="d-bgc-magenta-100 d-py8 d-px16 d-bar8 d-w216 lg:d-w332">
     <p class="d-fs-200 d-ww-inherit">Here's an example sentence to show how word-wrap works. Thisisasignlewordtodenotethedifferencebetweenthedifferentwaytowrapaword. Vivamus ullamcorperatduiaultrices eu lobortis nulla, sed vulputate orci. 这是一个中文例句，以举例说明断字的工作方式。単語分割の動作の例を示す日本語のサンプル文は次のとおりです。다음은 단어 분리 작동 방식의 예를 제공하는 한국어 샘플 문장입니다.</p>
   </div>
 </code-well-header>
